Refactor: #463 토론 목록 uiState 변수명 변경 및 테스트 코드 삭제
- `allDiscussions`를 `latestDiscussions`로 변경하여 토론 목록의 의미를 명확히 함
- 변경된 변수명에 맞게 UI 코드 수정
- `DiscussionsViewModelTest.kt` 테스트 코드 삭제

+++ android/TodokTodok/app/src/test/java/com/team/todoktodok/fixture/LatestDiscussionFixture.kt
@@ -10,7 +10,7 @@ val LATEST_DISCUSSIONS_RESPONSE =
             author =
                 AuthorResponse(
                     email = "user$index@example.com",
-                    id = index,
+                    id = index.toLong(),
                     nickname = "User$index",
                     profileImage = "https://example.com/profiles/$index.png",
                 ),
@@ -24,7 +24,7 @@ val LATEST_DISCUSSIONS_RESPONSE =
             commentCount = (0..50).random(),
             content = "This is a sample discussion content for discussion #$index.",
             createdAt = "2025-08-20T10:59:48",
-            discussionId = index,
+            discussionId = index.toLong(),
             isLikedByMe = listOf(true, false).random(),
             likeCount = (0..100).random(),
             title = "Discussion Title $index",
